extracting methods to its own constructor
